requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
flaskwebgui==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psutil==5.9.8
PyYAML==6.0.1
Werkzeug==3.0.1
